+++ pkg/resource/grafana-agent-config/reconciler.go
@@ -3,10 +3,8 @@ package grafanaagentconfig
 import (
 	"context"
 	"reflect"
-	"time"
 
 	"github.com/blang/semver"
-	appv1 "github.com/giantswarm/apiextensions-application/api/v1alpha1"
 	"github.com/pkg/errors"
 	v1 "k8s.io/api/core/v1"
 	apimachineryerrors "k8s.io/apimachinery/pkg/api/errors"
@@ -31,14 +29,9 @@ func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Inter
 	logger := log.FromContext(ctx)
 	logger.Info("grafana-agent-config create")
 
-	observabilityBundleVersion, err := common.GetObservabilityBundleAppVersion(lc, r.Client, ctx)
-	if err != nil {
-		// Handle case where the app is not found.
-		if apimachineryerrors.IsNotFound(err) {
-			logger.Info("grafana-agent-config - observability bundle app not found, requeueing")
-			// If the app is not found we should requeue and try again later (5 minutes is the app platform default reconciliation time)
-			return ctrl.Result{RequeueAfter: time.Duration(5 * time.Minute)}, nil
-		}
+	observabilityBundleVersion, ok := common.ObservabilityBundleAppVersionFromContext(ctx)
+	if !ok {
+		err := errors.New("grafana-agent-config - observability bundle app version not found in context")
 		return ctrl.Result{}, errors.WithStack(err)
 	}
 
@@ -47,17 +40,10 @@ func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Inter
 		return ctrl.Result{}, nil
 	}
 
-	// Get observability bundle app metadata.
-	appMeta := common.ObservabilityBundleAppMeta(lc)
-	// Retrieve the app.
-	var currentApp appv1.App
-	err = r.Client.Get(ctx, types.NamespacedName{Name: lc.AppConfigName("grafana-agent"), Namespace: appMeta.GetNamespace()}, &currentApp)
-	if err != nil {
-		if apimachineryerrors.IsNotFound(err) {
-			logger.Info("grafana-agent-config - app not found, requeuing")
-			// If the app is not found we should requeue and try again later (5 minutes is the app platform default reconciliation time)
-			return ctrl.Result{RequeueAfter: time.Duration(5 * time.Minute)}, nil
-		}
+	// Retrieve the Grafana agent app.
+	currentApp, ok := common.GrafanaAgentAppFromContext(ctx)
+	if !ok {
+		err := errors.New("grafana-agent-config - grafana agent app not found in context")
 		return ctrl.Result{}, errors.WithStack(err)
 	}