+++ pkg/resource/logging-secret/reconciler.go
@@ -30,16 +30,16 @@ func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Inter
 	logger.Info("logging-secret create")
 
 	// Retrieve secret containing credentials
-	var loggingCredentialsSecret v1.Secret
-	err := r.Client.Get(ctx, types.NamespacedName{Name: loggingcredentials.LoggingCredentialsSecretMeta(lc).Name, Namespace: loggingcredentials.LoggingCredentialsSecretMeta(lc).Namespace},
-		&loggingCredentialsSecret)
-	if err != nil {
+	loggingCredentialsSecret, ok := loggingcredentials.FromContext(ctx)
+	if !ok {
+		err := errors.New("logging-secret - logging credentials secret not found in context")
 		return ctrl.Result{}, errors.WithStack(err)
 	}
 
 	// Retrieve Loki ingress name
-	lokiURL, err := common.ReadLokiIngressURL(ctx, lc, r.Client)
-	if err != nil {
+	lokiURL, ok := common.LokiIngressURLFromContext(ctx)
+	if !ok {
+		err := errors.New("loki ingress URL not found in context")
 		return ctrl.Result{}, errors.WithStack(err)
 	}