+++ pkg/resource/loki-auth/reconciler.go
@@ -29,10 +29,9 @@ func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Inter
 	logger.Info("lokiauth create")
 
 	// Retrieve secret containing credentials
-	var lokiAuthSecret v1.Secret
-	err := r.Client.Get(ctx, types.NamespacedName{Name: loggingcredentials.LoggingCredentialsSecretMeta(lc).Name, Namespace: loggingcredentials.LoggingCredentialsSecretMeta(lc).Namespace},
-		&lokiAuthSecret)
-	if err != nil {
+	lokiAuthSecret, ok := loggingcredentials.FromContext(ctx)
+	if !ok {
+		err := errors.New("lokiauth - logging credentials secret not found in context")
 		return ctrl.Result{}, errors.WithStack(err)
 	}